refactor(ios): Extract ContentView components into separate files
Comprehensive refactoring to improve code organization and maintainability.

**ContentView Reduction:**
- Before: 1110 lines
- After: 384 lines
- Reduction: 65% (726 lines)

**New Components Created:**

1. **Views/Components/EmptyStateView.swift** (124 lines)
   - Extracted empty state UI with permission-aware messaging
   - Includes all accessibility identifiers from UI tests
   - Clean parameter-based API for state injection
   - Reusable across different views

2. **Views/Components/ExpenseRowView.swift** (58 lines)
   - Standalone expense row component
   - Handles formatting and display logic
   - Includes voice indicator icon
   - Extracted from ContentView lines 989-1044

3. **Services/SpeechRecognitionManager.swift** (284 lines)
   - Manages all speech recognition logic
   - Handles audio engine lifecycle with fallback format support
   - Implements auto-stop detection with silence threshold
   - Provides callbacks for transcription results and errors
   - Published properties for UI state binding
   - Proper [weak self] and DispatchQueue.main.async usage

4. **Services/PermissionManager.swift** (146 lines)
   - Centralizes permission management
   - Handles speech recognition and microphone permissions
   - Manages permission request flow
   - Provides app settings navigation
   - **@MainActor annotation on handleInitialPermissionResults** (fixes concurrency)
   - Proper lifecycle manager integration

5. **SiriKit/AddToSiriDelegate.swift** (56 lines)
   - Extracted Siri shortcut delegate
   - Handles voice shortcut creation callbacks
   - Manages success/error notifications
   - Extracted from ContentView lines 1056-1109

**Benefits:**
- Improved code organization with clear separation of concerns
- Each component has single responsibility
- Easier testing and maintenance
- Better reusability of components
- Reduced complexity in ContentView
- Cleaner architecture following MVVM pattern
- All Swift concurrency requirements properly handled

**ContentView Responsibilities (Now):**
- View composition and state coordination
- Voice command processing
- Navigation between different view states
- Alert and notification handling
- Onboarding flow management

**Technical Notes:**
- All onboarding logic preserved (test flags, completion state)
- All accessibility identifiers maintained for UI tests
- Audio format validation with fallback support included
- Proper @MainActor usage for lifecycle manager access
- All files follow Swift naming conventions and patterns

**Based on:** develop (3152cdb)
**Branch created from:** Latest develop with CI/CD fixes and onboarding
